+++ .mochaarc.json
@@ -1,0 +1,8 @@
+{
+  "require": [
+    "test/helpers/init.js"
+  ],
+  "recursive": true,
+  "reporter": "spec",
+  "timeout": 5000
+}
