+++ test/lib/fauna-object-translator.test.js
@@ -1,0 +1,8 @@
+
+
+describe('FaunaObjectTranslator', () => {
+
+  it('should run a test', () => {
+    console.log('hello world')
+  })
+})
