--- test/mocha.opts
@@ -1,4 +1,0 @@
---require test/helpers/init.js
---recursive
---reporter spec
---timeout 5000
